UPSTREAM: regmap: Move tracing header into drivers/base/regmap
The tracing events for regmap are confined to the regmap subsystem. It
also requires accessing an internal header. Instead of including the
internal header from a generic file location, move the tracing file
into the regmap directory.

Also rename the regmap tracing header to trace.h, as it is redundant to
keep the regmap.h name when it is in the regmap directory.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit f58078dacaaf6fbe664d5404dca7f76dce576113)

BUG=chrome-os-partner:46758
TEST=`emerge-oak chromeos-kernel-3_18`

Change-Id: Ia74d6ffab341729363f3db8a1e1487b4c3049ed9
Signed-off-by: Andrew-CT Chen <andrew-ct.chen@mediatek.com>
Reviewed-on: https://chromium-review.googlesource.com/307757
Commit-Ready: Wei-Ning Huang <wnhuang@chromium.org>
Tested-by: Wei-Ning Huang <wnhuang@chromium.org>
Reviewed-by: Wei-Ning Huang <wnhuang@chromium.org>