UPSTREAM: PM / domains: factor out code to get the generic PM domain from a struct device
The PM domain code contains two methods to get the generic PM domain
for a struct device.  One is dev_to_genpd() which is only safe when
we know for certain that the device has a generic PM domain attached.
The other is coded into genpd_dev_pm_detach() which ensures that the
PM domain in the struct device is a generic PM domain (and so is safer).

This commit factors out the safer version, documents it, and hides the
unsafe dev_to_genpd().

Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
(cherry picked from commit 446d999c1c92cec996e759dc3c03110596e626f5)
Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chromium:571567
TEST=build & boot to UI on oak-rev4
TEST=build & CQ test smaug, strago

Change-Id: If9fc2592041aeead2e57c0f440fa9e3a352c107e
Reviewed-on: https://chromium-review.googlesource.com/319762
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>