UPSTREAM: ipvlan: fix addr hash list corruption
When ipvlan interface with IP addresses attached is brought down and then
deleted, the assigned addresses are deleted twice from the address hash
list, first on the interface down and second on the link deletion.
Similarly, when an address is added while the interface is down, it is added
second time once the interface is brought up.

When the interface is down, the addresses should be kept off the hash list
for performance reasons. Ensure this is true, which also fixes the double add
problem. To fix the double free, check whether the address is hashed before
removing it.

Reported-by: Dan Williams <dcbw@redhat.com>
Signed-off-by: Jiri Benc <jbenc@redhat.com>
Signed-off-by: Mahesh Bandewar <maheshb@google.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 27705f7085ce2e124fac4c280ce824962cc90bb6)
Signed-off-by: Amey Deshpande <ameyd@google.com>

BUG=b:21611043
TEST=None

Change-Id: Ib243d52b6c9cb29ec8b748e1f79fa4060b00d813
Reviewed-on: https://chromium-review.googlesource.com/276484
Reviewed-by: Filipe Brandenburger <filbranden@chromium.org>
Tested-by: Amey Deshpande <ameyd@google.com>
Commit-Queue: Amey Deshpande <ameyd@google.com>