UPSTREAM: drm/tegra: sor: Reset during initialization
As there isn't a way for the firmware on the Nyan Chromebooks to hand
over the display to the kernel, and the kernel isn't redoing the whole
configuration at present.

With this patch, the SOR is brought to a known state and we get correct
display on every boot.

Signed-off-by: Tomeu Vizoso <tomeu.vizoso@collabora.com>
Signed-off-by: Thierry Reding <treding@nvidia.com>
(cherry picked from commit 535a65db484ff4e4654a037f2ea7e1ff21431b77)
Signed-off-by: Stéphane Marchesin <marcheu@chromium.org>

Change-Id: I52c9bd281eff2b21eecc8b15ac8fbf25d2ea53d5
Reviewed-on: https://chromium-review.googlesource.com/267227
Tested-by: David Riley <davidriley@chromium.org>
Reviewed-by: David Riley <davidriley@chromium.org>
Commit-Queue: David Riley <davidriley@chromium.org>