UPSTREAM: ALSA: hda/generic - Make snd_hda_gen_path_power_filter() always applicable
Add the check of power_save_node flag at the beginning of the function
so that it skips the rest if the flag isn't set.  In this way, we can
call this function safely no matter whether the widget power-saving is
really used or not.

Signed-off-by: Takashi Iwai <tiwai@suse.de>
Link: git://git.kernel.org/pub/scm/linux/kernel/git/next/linux-next.git
br: master
(cherry picked from commit b6c09b3c7b54743cbc58af8c27bbae8299c4f8cc)
Signed-off-by: Vedang Patel <vedang.patel@intel.com>

BUG=None
TEST=with the whole series applied, the audio subsystem is stable.
TEST=there are no regressions observed on BSW.

Change-Id: I75f09ed32e3541783b0f045efa3dd8440b29a110