CHROMIUM: clocksource: tegra: add Tegra210 timer support
Add support for the Tegra210 timer that runs at oscillator clock.
We need this timer to work as system timer to replace the ARMv8
architected timer due to it can't survive across the power cycle of the
CPU core or CPUPORESET signal. So it can't be a wake up source when CPU
suspend in power down state.

Based-on-the-original-work-by:
Antti P Miettinen <amiettinen@nvidia.com>

BUG=chrome-os-partner:40741
TEST=boot the device and make sure it can be used as system timer
     cat /proc/interrupts
     CPU0 CPU1 CPU2 CPU3
     ...
     1: 10593 0 0 0 GIC 208 tegra210_timer0
     2: 0 5180 0 0 GIC 209 tegra210_timer1
     3: 0 0 3413 0 GIC 210 tegra210_timer2
     4: 0 0 0 3498 GIC 211 tegra210_timer3

Change-Id: I4c200d97867678677afa502ebaffb9708b82b749
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/269273
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>